feat: improve circular dependencies insight output and move GraphQL tests
